moved some html strings into own modules
