+++ src/utils/generateHtml.js
@@ -2,10 +2,15 @@
 const { capitalCase } = require("change-case");
 
 //requiring classes
-const Employee = require("../lib/Employee");
 const Manager = require("../lib/Manager");
 const Engineer = require("../lib/Engineer");
 const Intern = require("../lib/Intern");
+const createHeadHtml = require("./createHeadHtml");
+const {
+  createManagerCardHtml,
+  createEngineerCardHtml,
+  createInternCardHtml,
+} = require("./createCardsHtml");
 
 //main function to generate the overall html string
 const generateHtml = (team) => {
@@ -25,32 +30,6 @@ const generateHtml = (team) => {
   };
   console.table(recapTable);
 
-  //static string for html head section
-  const headHtml = `<head>
-  <meta charset="utf-8" />
-  <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-  <title>Team Profile</title>
-  <meta name="description" content="" />
-  <meta name="viewport" content="width=device-width, initial-scale=1" />
-
-  <link
-    rel="stylesheet"
-    href="https://cdnjs.cloudflare.com/ajax/libs/bootstrap/4.6.1/css/bootstrap.min.css"
-  />
-  <link
-    rel="stylesheet"
-    href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.1.1/css/all.min.css"
-    crossorigin="anonymous"
-  />
-  <link rel="preconnect" href="https://fonts.googleapis.com" />
-  <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-  <link
-    href="https://fonts.googleapis.com/css2?family=Roboto:wght@400;700&display=swap"
-    rel="stylesheet"
-  />
-  <link rel="stylesheet" href="./assets/css/styles.css" />
-</head>`;
-
   //create string for header
   const headerHtml = `<header class="jumbotron jumbotron-fluid header-title mb-0">
   <div class="container">
@@ -60,33 +39,6 @@ const generateHtml = (team) => {
 
   //create string for manager section
   const createManagerHtml = () => {
-    const createManagerCardHtml = (each) => {
-      return `<div class="card employee-card manager-card m-2">
-      <div class="card-body bg-manager">
-        <h5 class="card-title">${capitalCase(each.name)}</h5>
-        <p class="card-subtitle">
-          <i class="fa-solid fa-people-roof"></i> Manager
-        </p>
-      </div>
-      <ul class="list-group list-group-flush p-2 list-container">
-        <li class="list-group-item">ID: ${each.id}</li>
-        <li class="list-group-item d-flex flex-row">
-          <p class="link-text pr-2">
-            <i class="fa-solid fa-envelope"></i> Email:
-          </p>
-          <a href="mailto:${each.email}" target="_blank" class="card-link"
-            >${each.email}</a
-          >
-        </li>
-        <li class="list-group-item">
-          <i class="fa-solid fa-door-closed"></i> Office number: ${
-            each.officeNumber
-          }
-        </li>
-      </ul>
-    </div>`;
-    };
-
     return `<section
         class="manager-container d-flex flex-column align-items-center separator"
         id="manager-section">
@@ -96,39 +48,6 @@ const generateHtml = (team) => {
 
   //create string for engineer section
   const createEngineerHtml = () => {
-    const createEngineerCardHtml = (each) => {
-      return `<div class="card employee-card engineer-card m-2">
-    <div class="card-body bg-engineer">
-      <h5 class="card-title">${capitalCase(each.name)}</h5>
-      <p class="card-subtitle">
-        <i class="fa-solid fa-brain"></i> Engineer
-      </p>
-    </div>
-    <ul class="list-group list-group-flush p-2 list-container">
-      <li class="list-group-item">ID: ${each.id}</li>
-      <li class="list-group-item d-flex flex-row">
-        <p class="link-text pr-2">
-          <i class="fa-solid fa-envelope"></i> Email:
-        </p>
-        <a href="mailto:${each.email}" target="_blank" class="card-link"
-          >${each.email}</a
-        >
-      </li>
-      <li class="list-group-item d-flex flex-row">
-        <p class="link-text pr-2">
-          <i class="fa-brands fa-github"></i> Github:
-        </p>
-        <a
-          href="https://github.com/${each.githubUsername}"
-          target="_blank"
-          class="card-link"
-          >${each.githubUsername}</a
-        >
-      </li>
-    </ul>
-    </div>`;
-    };
-
     return engineers.length === 0
       ? `<section
       class="engineer-container d-flex flex-row flex-wrap justify-content-around align-items-center separator"
@@ -144,37 +63,6 @@ const generateHtml = (team) => {
 
   //create string for intern section
   const createInternHtml = () => {
-    const createInternCardHtml = (each) => {
-      return `<div class="card employee-card intern-card m-2">
-          <div class="card-body bg-intern">
-            <h5 class="card-title">${capitalCase(each.name)}</h5>
-            <p class="card-subtitle">
-              <i class="fa-solid fa-seedling"></i> Intern
-            </p>
-          </div>
-          <ul class="list-group list-group-flush p-2 list-container">
-            <li class="list-group-item">ID: ${each.id}</li>
-            <li class="list-group-item d-flex flex-row">
-              <p class="link-text pr-2">
-                <i class="fa-solid fa-envelope"></i> Email:
-              </p>
-              <a
-                href="mailto:${each.email}"
-                target="_blank"
-                class="card-link"
-                >${each.email}</a
-              >
-            </li>
-            <li class="list-group-item d-flex flex-row">
-              <p class="link-text pr-2">
-                <i class="fa-solid fa-graduation-cap"></i> School:
-              </p>
-              <p class="link-text">${each.school}</p>
-            </li>
-          </ul>
-          </div>`;
-    };
-
     return interns.length === 0
       ? `<section
       class="intern-container d-flex flex-row flex-wrap justify-content-around align-items-center"
@@ -194,6 +82,7 @@ const generateHtml = (team) => {
 </footer>`;
 
   //call to generate the section strings
+  const headHtml = createHeadHtml("TeamProfile");
   const managerHtml = createManagerHtml();
   const engineerHtml = createEngineerHtml();
   const internHtml = createInternHtml();
